+++ example/lib/settings/settings_model.dart
@@ -2,10 +2,11 @@
 
 import 'package:flutter/foundation.dart';
 
-final topLevelNavigationType = ValueNotifier<TopLevelNavigationType>(TopLevelNavigationType.bottom);
+final topLevelNavigationType = ValueNotifier<TopLevelNavigationType>(TopLevelNavigationType.bottomMaterial3);
 
 enum TopLevelNavigationType {
   bottom,
+  bottomMaterial3,
   drawer,
   tabs,
 }