+++ lib/src/navigation_scheme.dart
@@ -123,7 +123,7 @@ class NavigationScheme with ChangeNotifier {
   /// Stores the original destination in case of redirection.
   ///
   Destination? get redirectedFrom =>
-      _currentDestination.configuration.redirectedFrom;
+      _currentDestination.settings.redirectedFrom;
 
   bool _shouldClose = false;
 
@@ -167,7 +167,7 @@ class NavigationScheme with ChangeNotifier {
       return SynchronousFuture(null);
     }
     Log.d(runtimeType,
-        'goTo(): navigator=${navigator.tag}, destination=$destination, redirectedFrom=${destination.configuration.redirectedFrom}');
+        'goTo(): navigator=${navigator.tag}, destination=$destination, redirectedFrom=${destination.settings.redirectedFrom}');
     _shouldClose = false;
 
     final completer = Completer<void>();
@@ -232,7 +232,7 @@ class NavigationScheme with ChangeNotifier {
       notifyListeners();
       return;
     }
-    goTo(resolvedDestination.withConfiguration(resolvedDestination.configuration
+    goTo(resolvedDestination.withSettings(resolvedDestination.settings
         .copyWith(redirectedFrom: requestedDestination)));
   }
 
@@ -256,8 +256,8 @@ class NavigationScheme with ChangeNotifier {
 
   void _handleError(Destination? destination) {
     if (errorDestination != null) {
-      goTo((errorDestination!).withConfiguration(errorDestination!.configuration
-          .copyWith(redirectedFrom: destination)));
+      goTo((errorDestination!).withSettings(
+          errorDestination!.settings.copyWith(redirectedFrom: destination)));
     } else {
       throw UnknownDestinationException(destination);
     }
@@ -290,9 +290,8 @@ class NavigationScheme with ChangeNotifier {
           _updateCurrentDestination(backFrom: navigator.backFrom);
         }
       } else {
-        if (navigator.currentDestination.configuration.reset) {
-          goTo(owner
-              .withConfiguration(owner.configuration.copyWith(reset: true)));
+        if (navigator.currentDestination.settings.reset) {
+          goTo(owner.withSettings(owner.settings.copyWith(reset: true)));
         } else {
           goTo(owner);
         }
@@ -322,9 +321,9 @@ class NavigationScheme with ChangeNotifier {
     Log.d(runtimeType,
         'updateCurrentDestination(): currentDestination=$_currentDestination, newDestination=$newDestination');
     if (_currentDestination != newDestination ||
-        newDestination.configuration.reset) {
+        newDestination.settings.reset) {
       _currentDestination = newDestination;
-      if (_currentDestination == backFrom?.configuration.redirectedFrom) {
+      if (_currentDestination == backFrom?.settings.redirectedFrom) {
         notifyListeners();
         return;
       }
@@ -359,8 +358,7 @@ class NavigationScheme with ChangeNotifier {
           true)) {
         _destinationCompleters[destinationToComplete]?.complete();
       }
-      destinationToComplete =
-          destinationToComplete.configuration.redirectedFrom;
+      destinationToComplete = destinationToComplete.settings.redirectedFrom;
     }
   }
 }