+++ lib/src/widgets/bottom_navigation_builder.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 
+import '../destination.dart';
 import '../navigation_controller.dart';
 import 'index.dart';
 
@@ -25,9 +26,26 @@ class BottomNavigationBuilder implements NavigatorBuilder {
   /// Creates a [BottomNavigationBuilder] instance.
   ///
   const BottomNavigationBuilder({
-    required this.bottomNavigationItems,
+    this.bottomNavigationItems = const <BottomNavigationBarItem>[],
     this.parameters = const BottomNavigationBarParameters(),
-  });
+    this.navigationBarItems = const <NavigationDestination>[],
+    this.navigationBarParameters = const NavigationBarParameters(),
+    bool? material3,
+  }) : _material3 = material3 ?? false;
+
+  /// Creates a [BottomNavigationBuilder] instance that uses Material 3 [NavigationBar]
+  /// widget.
+  ///
+  factory BottomNavigationBuilder.navigationBar({
+    required List<NavigationDestination> navigationBarItems,
+    NavigationBarParameters? navigationBarParameters,
+  }) =>
+      BottomNavigationBuilder(
+        navigationBarItems: navigationBarItems,
+        navigationBarParameters:
+            navigationBarParameters ?? const NavigationBarParameters(),
+        material3: true,
+      );
 
   /// A list of [BottomNavigationBarItems], that corresponds to the navigator's
   /// destination list.
@@ -45,17 +63,37 @@ class BottomNavigationBuilder implements NavigatorBuilder {
   ///
   final BottomNavigationBarParameters parameters;
 
+  /// A list of [NavigationDestination] widgets, that corresponds to the navigator's
+  /// destination list.
+  ///
+  /// The list must contain the same number of items, following with the same order
+  /// as a destination list specified in the navigator.
+  ///
+  final List<NavigationDestination> navigationBarItems;
+
+  /// A set of [NavigationBar] parameters.
+  ///
+  /// Contains all supported parameters to customize [NavigationBar] widget.
+  /// Doesn't include 'items', 'onTap' and 'currentIndex', which are managed by
+  /// [BottomNavigationBuilder].
+  ///
+  final NavigationBarParameters navigationBarParameters;
+
+  final bool _material3;
+
   @override
   Widget build(BuildContext context, NavigationController navigator) {
     final currentDestination = navigator.currentDestination;
-    final content = currentDestination.build(context);
     return _BottomNavigationWrapper(
-      content: content,
-      items: bottomNavigationItems,
-      parameters: parameters,
+      destination: currentDestination,
       onSelectBottomTab: (index) =>
           navigator.goTo(navigator.destinations[index]),
       selectedIndex: navigator.destinations.indexOf(currentDestination),
+      items: bottomNavigationItems,
+      parameters: parameters,
+      navigationBarItems: navigationBarItems,
+      navigationBarParameters: navigationBarParameters,
+      material3: _material3,
     );
   }
 }
@@ -63,14 +101,17 @@ class BottomNavigationBuilder implements NavigatorBuilder {
 class _BottomNavigationWrapper extends StatefulWidget {
   const _BottomNavigationWrapper({
     Key? key,
-    required this.content,
-    required this.items,
+    required this.destination,
     required this.onSelectBottomTab,
     required this.selectedIndex,
-    required this.parameters,
+    this.items = const <BottomNavigationBarItem>[],
+    this.parameters = const BottomNavigationBarParameters(),
+    this.navigationBarItems = const <NavigationDestination>[],
+    this.navigationBarParameters = const NavigationBarParameters(),
+    this.material3 = false,
   }) : super(key: key);
 
-  final Widget content;
+  final Destination destination;
 
   final List<BottomNavigationBarItem> items;
 
@@ -80,43 +121,76 @@ class _BottomNavigationWrapper extends StatefulWidget {
 
   final BottomNavigationBarParameters parameters;
 
+  final List<NavigationDestination> navigationBarItems;
+
+  final NavigationBarParameters navigationBarParameters;
+
+  final bool material3;
+
   @override
   _BottomNavigationWrapperState createState() =>
       _BottomNavigationWrapperState();
 }
 
 class _BottomNavigationWrapperState extends State<_BottomNavigationWrapper> {
+  final _content = <Destination, Widget>{};
+
+  final _indexes = <Destination, int>{};
+
   late final OverlayEntry _mainOverlay;
 
   @override
   void initState() {
     super.initState();
+    _content[widget.destination] = widget.destination.build(context);
+    _indexes[widget.destination] = widget.selectedIndex;
     _mainOverlay = OverlayEntry(
       builder: (context) => Scaffold(
-        body: widget.content,
-        bottomNavigationBar: BottomNavigationBar(
-          items: widget.items,
-          currentIndex: widget.selectedIndex,
-          onTap: widget.onSelectBottomTab,
-          elevation: widget.parameters.elevation,
-          type: widget.parameters.type,
-          fixedColor: widget.parameters.fixedColor,
-          backgroundColor: widget.parameters.backgroundColor,
-          iconSize: widget.parameters.iconSize,
-          selectedItemColor: widget.parameters.selectedItemColor,
-          unselectedItemColor: widget.parameters.unselectedItemColor,
-          selectedIconTheme: widget.parameters.selectedIconTheme,
-          unselectedIconTheme: widget.parameters.unselectedIconTheme,
-          selectedFontSize: widget.parameters.selectedFontSize,
-          unselectedFontSize: widget.parameters.unselectedFontSize,
-          selectedLabelStyle: widget.parameters.selectedLabelStyle,
-          unselectedLabelStyle: widget.parameters.unselectedLabelStyle,
-          showSelectedLabels: widget.parameters.showSelectedLabels,
-          showUnselectedLabels: widget.parameters.showUnselectedLabels,
-          mouseCursor: widget.parameters.mouseCursor,
-          enableFeedback: widget.parameters.enableFeedback,
-          landscapeLayout: widget.parameters.landscapeLayout,
+        body: Stack(
+          children: [
+            ..._content.entries
+                .map((entry) => Offstage(
+                      offstage: _indexes[entry.key] != widget.selectedIndex,
+                      child: entry.value,
+                    ))
+                .toList(),
+          ],
         ),
+        bottomNavigationBar: widget.material3
+            ? NavigationBar(
+                animationDuration:
+                    widget.navigationBarParameters.animationDuration,
+                selectedIndex: widget.selectedIndex,
+                destinations: widget.navigationBarItems,
+                onDestinationSelected: widget.onSelectBottomTab,
+                backgroundColor: widget.navigationBarParameters.backgroundColor,
+                elevation: widget.navigationBarParameters.elevation,
+                height: widget.navigationBarParameters.height,
+                labelBehavior: widget.navigationBarParameters.labelBehavior,
+              )
+            : BottomNavigationBar(
+                items: widget.items,
+                currentIndex: widget.selectedIndex,
+                onTap: widget.onSelectBottomTab,
+                elevation: widget.parameters.elevation,
+                type: widget.parameters.type,
+                fixedColor: widget.parameters.fixedColor,
+                backgroundColor: widget.parameters.backgroundColor,
+                iconSize: widget.parameters.iconSize,
+                selectedItemColor: widget.parameters.selectedItemColor,
+                unselectedItemColor: widget.parameters.unselectedItemColor,
+                selectedIconTheme: widget.parameters.selectedIconTheme,
+                unselectedIconTheme: widget.parameters.unselectedIconTheme,
+                selectedFontSize: widget.parameters.selectedFontSize,
+                unselectedFontSize: widget.parameters.unselectedFontSize,
+                selectedLabelStyle: widget.parameters.selectedLabelStyle,
+                unselectedLabelStyle: widget.parameters.unselectedLabelStyle,
+                showSelectedLabels: widget.parameters.showSelectedLabels,
+                showUnselectedLabels: widget.parameters.showUnselectedLabels,
+                mouseCursor: widget.parameters.mouseCursor,
+                enableFeedback: widget.parameters.enableFeedback,
+                landscapeLayout: widget.parameters.landscapeLayout,
+              ),
       ),
     );
   }
@@ -124,7 +198,21 @@ class _BottomNavigationWrapperState extends State<_BottomNavigationWrapper> {
   @override
   void didUpdateWidget(_BottomNavigationWrapper oldWidget) {
     super.didUpdateWidget(oldWidget);
-    if (oldWidget.selectedIndex != widget.selectedIndex) {
+    bool needsRebuild = false;
+    if (widget.material3 != oldWidget.material3) {
+      needsRebuild = true;
+    }
+    if (!widget.destination.isFinalDestination) {
+      needsRebuild = true;
+      _content[widget.destination] = widget.destination.build(context);
+      _indexes[widget.destination] = widget.selectedIndex;
+    } else if (oldWidget.selectedIndex != widget.selectedIndex &&
+        !_content.containsKey(widget.destination)) {
+      needsRebuild = true;
+      _content[widget.destination] = widget.destination.build(context);
+      _indexes[widget.destination] = widget.selectedIndex;
+    }
+    if (needsRebuild) {
       _mainOverlay.markNeedsBuild();
     }
   }
@@ -243,4 +331,45 @@ class BottomNavigationBarParameters {
   /// [BottomNavigationBar.landscapeLayout]
   ///
   final BottomNavigationBarLandscapeLayout? landscapeLayout;
+}
+
+/// Contains parameters to customize the [NavigationBar].
+///
+/// It includes all the same arguments as the [NavigationBar()], excepting
+/// the 'items', 'onTap' and 'currentIndex', which are managed by the [BottomNavigationBuilder].
+///
+/// See also:
+/// - [BottomNavigationBuilder]
+/// - [NavigationBar]
+///
+class NavigationBarParameters {
+  /// Create a [NavigationBarParameters] instance.
+  ///
+  const NavigationBarParameters({
+    this.animationDuration,
+    this.backgroundColor,
+    this.elevation,
+    this.height,
+    this.labelBehavior,
+  });
+
+  /// [NavigationBar.animationDuration]
+  ///
+  final Duration? animationDuration;
+
+  /// [NavigationBar.backgroundColor]
+  ///
+  final Color? backgroundColor;
+
+  /// [NavigationBar.elevation]
+  ///
+  final double? elevation;
+
+  /// [NavigationBar.height]
+  ///
+  final double? height;
+
+  /// [NavigationBar.labelBehavior]
+  ///
+  final NavigationDestinationLabelBehavior? labelBehavior;
 }